backend/requirements.txt
# Flask Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Environment
python-dotenv==1.0.0

# Security
Werkzeug==3.0.1

# Validation
email-validator==2.1.0

# Development
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
black==23.12.1
flake8==6.1.0
python-decouple==3.8

# Production
gunicorn==21.2.0
